+++ lib/bookshop/generators/bookshop/app/templates/config/config.rb
@@ -1,0 +1,4 @@
+Bookshop.configure do |config|
+  config.epubcheck = '/usr/bin/epubcheck'
+  config.kindlegen = '/usr/bin/kindlegen'
+end